build: Set HAS_STACK=1 when ipv6 is enabled
The variable HAS_STACK is set when ipv4 or rime stacks are enabled but
not when ipv6 is enabled. This patch sets HAS_STACK when ipv6 is enabled
so platform Makefiles can rely on it to find out if any network stack is
enabled.

This change enabled us to build networking-specific components (e.g.
radio driver) only if we network feature is enabled.

Signed-off-by: Andre Guedes <andre.guedes@intel.com>